Move html module to dataframe module
